fix: resolve compilation errors across canvas files
- CanvasMinimap.kt: capture MaterialTheme.colorScheme.primary before Canvas DrawScope (non-composable lambda)
- InfiniteCanvas.kt: add missing dp import, fix SubcomposeLayout .first().measure(constraints), store touch Modifier in variable
- LinkToEntityDialog.kt: fix Triple destructuring (Triple has 3 components, not 4) by splitting icon extraction
- CanvasBlock.kt: add BoxScope import for align() extension
- StickyNoteBlock.kt: fix hex literals exceeding Int.MAX_VALUE by using .toInt() to preserve ARGB values
- TextBlock.kt: add missing imports for TransformedText and Color

+++ app/src/main/java/fieldmind/research/app/features/field/presentation/canvas/blocks/StickyNoteBlock.kt
@@ -72,12 +72,12 @@ fun StickyNoteBlock(
     // ── Sticky note colors ──
     val colors = remember {
         listOf(
-            0xFFFFF9C4 to "Yellow",   // default
-            0xFFC8E6C9 to "Green",
-            0xFFF8BBD0 to "Pink",
-            0xFFBBDEFB to "Blue",
-            0xFFFFE0B2 to "Orange",
-            0xFFE1BEE7 to "Purple"
+            0xFFFFF9C4.toInt() to "Yellow",   // default
+            0xFFC8E6C9.toInt() to "Green",
+            0xFFF8BBD0.toInt() to "Pink",
+            0xFFBBDEFB.toInt() to "Blue",
+            0xFFFFE0B2.toInt() to "Orange",
+            0xFFE1BEE7.toInt() to "Purple"
         )
     }
